Add product detail template and functionality

+++ products/views.py
@@ -167,11 +167,20 @@ def update_product(request, product_id):
 @login_required
 def delete_product(request, product_id):
     """ Delete a product from the store """
+    product = get_object_or_404(Product, pk=product_id)
     if not request.user.is_superuser:
         messages.error(request, 'Sorry, only store owners can do that.')
         return redirect(reverse('home'))
 
-    product = get_object_or_404(Product, pk=product_id)
-    product.delete()
-    messages.success(request, 'Product deleted!')
-    return redirect(reverse('products'))
+    if request.method == "POST":
+      
+        product.delete()
+        messages.success(request, 'Product deleted successfully')
+        return redirect(reverse('products'))
+
+    template = 'products/delete_product.html'
+
+    context = {
+        'product': product
+        }
+    return render(request, template, context)
